move all components from "Pages" to "components" remove redundant menu & footer from components
