+++ frontend/src/App.js
@@ -15,10 +15,6 @@ import MainPage from './containers/Main/MainPage';
 import routes from './constants/routes';
 import Talent from './containers/Talent/TalentHome';
 import Employer from './containers/Employer/EmployerPage';
-import PrivacyPolicy from './pages/PrivacyPolicy/PrivacyPolicy'
-import TermsOfUse from './pages/TermsOfService/TermsOfUse'
-import Support from './pages/Support/Support'
-import Blog from './pages/Blog/Blog'
 
 const store = createStore(reducer, composeWithDevTools(applyMiddleware(thunk)));
 
@@ -47,10 +43,6 @@ class App extends React.Component <Props, State> {
             <Switch>
               <Route path={routes.TALENT_HOME} component={Talent}/>
               <Route path={routes.EMPLOYERS_HOME} component={Employer}/>
-              <Route path={routes.PRIVACY_POLICY} component={PrivacyPolicy}/>
-              <Route path={routes.TERMS_OF_USE} component={TermsOfUse}/>
-              <Route path={routes.SUPPORT_CENTER} component={Support}/>
-              <Route path={routes.BLOG} component={Blog}/>
               <Route path={routes.HOME} component={MainPage}/>
             </Switch>
           </Router>
