+++ frontend/src/components/Main/About/About.css
@@ -2,7 +2,8 @@
 @import url("https://fonts.googleapis.com/css?family=Carter+One&display=swap");
 #about-banner {
   background-image: url("../../../assets/images/main_about_bg.png");
-  position: relative; }
+  position: relative;
+  font-family: "Monserrat", sans-serif; }
   #about-banner .imageFilter {
     position: absolute;
     background-color: rgba(0, 0, 0, 0.5);
@@ -18,46 +19,48 @@
     color: #f2f2f2;
     transform: translateZ(0); }
 
-#what-we-do .contentContainer,
-#mission .contentContainer {
-  width: 50%; }
-  #what-we-do .contentContainer .greenDash,
-  #mission .contentContainer .greenDash {
-    margin-bottom: 2em; }
-  #what-we-do .contentContainer .row,
-  #mission .contentContainer .row {
+#what-we-do,
+#mission {
+  font-family: "Monserrat", sans-serif; }
+  #what-we-do .contentContainer,
+  #mission .contentContainer {
+    width: 50%; }
+    #what-we-do .contentContainer .greenDash,
+    #mission .contentContainer .greenDash {
+      margin-bottom: 2em; }
+    #what-we-do .contentContainer .row,
+    #mission .contentContainer .row {
+      display: flex;
+      width: 100%; }
+      #what-we-do .contentContainer .row .iconContainer,
+      #mission .contentContainer .row .iconContainer {
+        width: 60px;
+        min-width: 60px; }
+        #what-we-do .contentContainer .row .iconContainer img,
+        #mission .contentContainer .row .iconContainer img {
+          max-height: 40px; }
+      #what-we-do .contentContainer .row .rowContent h4,
+      #mission .contentContainer .row .rowContent h4 {
+        font-size: 16px;
+        font-weight: 700;
+        font-family: "Monserrat", sans-serif; }
+      #what-we-do .contentContainer .row .rowContent p,
+      #mission .contentContainer .row .rowContent p {
+        padding: 1em 0 2em;
+        font-size: 16px;
+        font-family: "Monserrat", sans-serif; }
+  #what-we-do .imageContainer,
+  #mission .imageContainer {
+    width: 50%;
     display: flex;
-    width: 100%; }
-    #what-we-do .contentContainer .row .iconContainer,
-    #mission .contentContainer .row .iconContainer {
-      width: 60px;
-      min-width: 60px; }
-      #what-we-do .contentContainer .row .iconContainer img,
-      #mission .contentContainer .row .iconContainer img {
-        max-height: 40px; }
-    #what-we-do .contentContainer .row .rowContent h4,
-    #mission .contentContainer .row .rowContent h4 {
-      font-size: 16px;
-      font-weight: 700;
-      font-family: "Monserrat", sans-serif; }
-    #what-we-do .contentContainer .row .rowContent p,
-    #mission .contentContainer .row .rowContent p {
-      padding: 1em 0 2em;
-      font-size: 16px;
-      font-family: "Monserrat", sans-serif; }
-
-#what-we-do .imageContainer,
-#mission .imageContainer {
-  width: 50%;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  justify-content: flex-end; }
-  #what-we-do .imageContainer img,
-  #mission .imageContainer img {
-    height: auto;
-    max-width: 500px;
-    width: 100%; }
+    flex-direction: column;
+    align-items: flex-end;
+    justify-content: flex-end; }
+    #what-we-do .imageContainer img,
+    #mission .imageContainer img {
+      height: auto;
+      max-width: 500px;
+      width: 100%; }
 
 #numbers .container,
 #what-we-do .container,
@@ -66,12 +69,14 @@
   max-width: 1200px;
   margin: 3em auto 0;
   padding: 0 20px;
-  justify-content: space-between; }
+  justify-content: space-between;
+  font-family: "Monserrat", sans-serif; }
 
 #numbers {
   background-color: #f2f2f2;
   padding-top: 1em;
-  padding-bottom: 4em; }
+  padding-bottom: 4em;
+  font-family: "Monserrat", sans-serif; }
   #numbers .column {
     display: flex;
     flex-direction: column;
@@ -91,7 +96,8 @@
       text-align: center; }
 
 #mission {
-  background-color: #fff; }
+  background-color: #fff;
+  font-family: "Monserrat", sans-serif; }
   #mission .imageContainer {
     align-items: center; }
     #mission .imageContainer img {
@@ -101,7 +107,8 @@
 
 #new-jobs {
   background-color: #fafafa;
-  padding: 2em; }
+  padding: 2em;
+  font-family: "Monserrat", sans-serif; }
   #new-jobs h4 {
     text-align: center;
     font-size: 26px;
@@ -111,7 +118,8 @@
     display: block; }
 
 #team {
-  padding: 2em 0 4em; }
+  padding: 2em 0 4em;
+  font-family: "Monserrat", sans-serif; }
   #team .teamMembersContainer {
     display: flex;
     max-width: 1200px;
@@ -150,13 +158,15 @@
           margin: 10px; }
 
 #partners-about {
-  margin: 3em; }
+  margin: 3em;
+  font-family: "Monserrat", sans-serif; }
 
 #feedback {
   background-image: url("../../../assets/images/feedback_bg.png");
   background-size: cover;
   background-position: center;
-  position: relative; }
+  position: relative;
+  font-family: "Monserrat", sans-serif; }
   #feedback .content {
     padding: 2em; }
     #feedback .content .feedbackContainer {
@@ -218,25 +228,27 @@
         border-radius: 50%;
         overflow: hidden; }
 
-#contact .content {
-  max-width: 1200px;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  margin: 3em auto; }
-  #contact .content .leftCol, #contact .content .rightCol {
-    width: 40%; }
-    #contact .content .leftCol .headerBlock, #contact .content .rightCol .headerBlock {
-      display: inline-block; }
-  #contact .content .rightCol p {
-    margin: 3em 0;
-    font-size: 14px;
-    line-height: 20px;
-    font-family: "Monserrat", sans-serif;
-    color: #828282; }
-  #contact .content .rightCol .contactData {
-    margin: 0;
-    padding: 0;
-    font-size: 16px; }
-    #contact .content .rightCol .contactData span {
-      font-weight: 600; }
+#contact {
+  font-family: "Monserrat", sans-serif; }
+  #contact .content {
+    max-width: 1200px;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    margin: 3em auto; }
+    #contact .content .leftCol, #contact .content .rightCol {
+      width: 40%; }
+      #contact .content .leftCol .headerBlock, #contact .content .rightCol .headerBlock {
+        display: inline-block; }
+    #contact .content .rightCol p {
+      margin: 3em 0;
+      font-size: 14px;
+      line-height: 20px;
+      font-family: "Monserrat", sans-serif;
+      color: #828282; }
+    #contact .content .rightCol .contactData {
+      margin: 0;
+      padding: 0;
+      font-size: 16px; }
+      #contact .content .rightCol .contactData span {
+        font-weight: 600; }
