+++ frontend/src/components/Main/Blog/Blog.css
@@ -1,0 +1,142 @@
+.blog-header {
+  width: 100%;
+  height: 500px;
+  background-image: url("../../../assets/images/bg-blog.png");
+  background-repeat: no-repeat;
+  background-position: center;
+  background-size: cover;
+  position: relative; }
+  .blog-header__title {
+    width: 988px;
+    padding: 0 10px;
+    font-size: 48px;
+    font-weight: bold;
+    color: #fff;
+    z-index: 99999;
+    position: absolute;
+    text-transform: uppercase;
+    text-align: center;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    line-height: 50px; }
+  .blog-header__image-filter {
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: #060C2E;
+    opacity: 50%; }
+
+.blog-container {
+  width: 1200px;
+  margin: 0 auto;
+  padding: 0 10px;
+  display: flex;
+  margin-top: 87px; }
+  .blog-container .blog-content {
+    width: 70%;
+    margin-right: 70px; }
+  .blog-container .blog-item {
+    color: #828282;
+    margin-bottom: 67px; }
+    .blog-container .blog-item__description {
+      margin-bottom: 4px;
+      font-size: 12px;
+      text-transform: uppercase; }
+    .blog-container .blog-item__container {
+      display: flex;
+      margin-bottom: 16px; }
+    .blog-container .blog-item__header {
+      margin-bottom: 25px;
+      color: #4F4F4F;
+      font-size: 36px;
+      font-weight: 600; }
+    .blog-container .blog-item__text {
+      font-size: 24px;
+      line-height: 29px; }
+    .blog-container .blog-item__icon {
+      margin-right: 24px;
+      width: 180px;
+      height: 115px;
+      flex-shrink: 0;
+      background-size: cover;
+      background-repeat: no-repeat;
+      background-position: center;
+      background-image: url("../../../assets/images/rachel-moenning-570271-unsplash.png"); }
+    .blog-container .blog-item__link {
+      text-transform: uppercase;
+      text-decoration: underline;
+      font-size: 18px;
+      cursor: pointer;
+      color: #828282; }
+    .blog-container .blog-item__link:hover {
+      text-decoration: none; }
+  .blog-container .blog-sidebar {
+    color: #828282; }
+    .blog-container .blog-sidebar .search-form {
+      height: 44px; }
+      .blog-container .blog-sidebar .search-form .search-icon {
+        position: relative;
+        z-index: 10;
+        top: 13px;
+        left: 17px; }
+      .blog-container .blog-sidebar .search-form .search-input {
+        width: 306px;
+        height: 44px;
+        box-sizing: border-box;
+        padding-left: 44px;
+        border: 1px solid #BDBDBD;
+        border-radius: 8px;
+        position: absolute; }
+        .blog-container .blog-sidebar .search-form .search-input::placeholder {
+          color: #828282; }
+    .blog-container .blog-sidebar .posts-title {
+      font-size: 24px;
+      color: #333333;
+      font-weight: 600;
+      text-transform: uppercase;
+      margin: 57px 0 28px; }
+      .blog-container .blog-sidebar .posts-title::after {
+        content: '';
+        display: block;
+        width: 306px;
+        height: 1px;
+        margin-top: 17px;
+        background: #BDBDBD; }
+    .blog-container .blog-sidebar .posts-list__item {
+      font-size: 18px;
+      margin-top: 10px;
+      line-height: 22px; }
+      .blog-container .blog-sidebar .posts-list__item::after {
+        content: '';
+        display: block;
+        width: 306px;
+        height: 1px;
+        margin-top: 17px;
+        background: #BDBDBD; }
+    .blog-container .blog-sidebar .form-title {
+      font-size: 18px;
+      color: #333333;
+      font-weight: 600;
+      margin-top: 100px; }
+  .blog-container .subscribe-form .mail-input {
+    width: 306px;
+    height: 44px;
+    box-sizing: border-box;
+    padding-left: 44px;
+    border: 1px solid #BDBDBD;
+    border-radius: 8px;
+    margin-top: 16px; }
+  .blog-container .subscribe-form .send-btn {
+    padding: 15px 30px;
+    display: block;
+    background: #FF6F60;
+    border-radius: 8px;
+    border: none;
+    color: #ffffff;
+    font-size: 20px;
+    font-weight: 500;
+    text-transform: uppercase;
+    margin-top: 20px; }
