+++ frontend/src/components/Main/Blog/Blog.scss
@@ -1,7 +1,7 @@
 .blog-header {
   width: 100%;
   height: 500px;
-  background-image: url("../../assets/images/bg-blog.png");
+  background-image: url("../../../assets/images/bg-blog.png");
   background-repeat: no-repeat;
   background-position: center;
   background-size: cover;
@@ -81,7 +81,7 @@
       background-size: cover;
       background-repeat: no-repeat;
       background-position: center;
-      background-image: url("../../assets/images/rachel-moenning-570271-unsplash.png");
+      background-image: url("../../../assets/images/rachel-moenning-570271-unsplash.png");
     }
 
     &__link {
