+++ frontend/src/components/Main/PrivacyPolicy/PrivacyPolicy.css
@@ -1,0 +1,49 @@
+@import url("https://fonts.googleapis.com/css?family=Montserrat:400,600,700&display=swap");
+@import url("https://fonts.googleapis.com/css?family=Carter+One&display=swap");
+.privacy-policy {
+  padding-top: 70px; }
+  .privacy-policy .privacy-policy-container {
+    width: 1200px;
+    min-height: calc(100vh - 170px);
+    margin: 0 auto;
+    padding: 0 30px; }
+    .privacy-policy .privacy-policy-container__header {
+      font-size: 32px;
+      font-weight: 600;
+      color: #4f4f4f;
+      text-transform: uppercase;
+      padding: 50px 0; }
+  .privacy-policy .privacy-policy-content {
+    display: flex;
+    font-size: 16px;
+    line-height: 21px; }
+    .privacy-policy .privacy-policy-content .privacy-policy-text {
+      width: 70%; }
+      .privacy-policy .privacy-policy-content .privacy-policy-text .sub-header {
+        font-size: 28px;
+        color: #333333;
+        margin-bottom: 10px; }
+    .privacy-policy .privacy-policy-content p {
+      margin-bottom: 30px; }
+    .privacy-policy .privacy-policy-content p:last-child {
+      margin-bottom: 100px; }
+    .privacy-policy .privacy-policy-content .side-menu {
+      width: 30%; }
+      .privacy-policy .privacy-policy-content .side-menu__item {
+        font-size: 18px;
+        cursor: pointer; }
+      .privacy-policy .privacy-policy-content .side-menu__item:hover:after {
+        background: rgba(54, 179, 168, 0.9);
+        height: 2px;
+        margin-bottom: 10px; }
+      .privacy-policy .privacy-policy-content .side-menu__item::after {
+        content: '';
+        display: block;
+        height: 1px;
+        width: 60%;
+        margin: 11px 0;
+        background-color: #BDBDBD; }
+  .privacy-policy .item-active > .side-menu__item::after {
+    background: rgba(54, 179, 168, 0.9);
+    height: 2px;
+    margin-bottom: 10px; }
