+++ frontend/src/components/Main/TermsOfService/TermsOfUse.js
@@ -1,16 +1,13 @@
 import React from 'react';
-import Footer from '../../components/Main/Footer/Footer'
-import TopMenuGrey from '../../components/TopMenuGray/TopMenuGrey'
-import { Link, animateScroll as scroll } from "react-scroll";
-import './PrivacyPolicy.scss'
+import {Link} from 'react-scroll';
+import './TermsOfUse.scss';
 
-function PrivacyPolicy(props) {
+function TermsOfUse(props) {
     return (
-        <div className='privacy-policy'>
-            <TopMenuGrey/>
-            <div className='privacy-policy-container'>
-                <h1 className='privacy-policy-container__header'>Privacy Policy</h1>
-                <div className='privacy-policy-content'>
+        <div className='terms-of-use'>
+            <div className='terms-of-use__container'>
+                <h1 className='terms-of-use__header'>Terms of service</h1>
+                <div className='terms-of-use__content'>
                     <div className='side-menu'>
                         <ul className='side-menu__nav'>
                             <Link
@@ -20,7 +17,7 @@ function PrivacyPolicy(props) {
                                 smooth={true}
                                 duration={1000}
                             >
-                                <li className="side-menu__item">Privacy Statement and Notice</li>
+                                <li className="side-menu__item">Introduction</li>
                             </Link>
                             <Link
                                 activeClass='item-active'
@@ -29,7 +26,7 @@ function PrivacyPolicy(props) {
                                 smooth={true}
                                 duration={1000}
                             >
-                                <li className="side-menu__item">What information we collect</li>
+                                <li className="side-menu__item">Services</li>
                             </Link>
                             <Link
                                 activeClass='item-active'
@@ -38,7 +35,7 @@ function PrivacyPolicy(props) {
                                 smooth={true}
                                 duration={1000}
                             >
-                                <li className="side-menu__item">Sharing your information</li>
+                                <li className="side-menu__item">Registration</li>
                             </Link>
                             <Link
                                 activeClass='item-active'
@@ -47,7 +44,7 @@ function PrivacyPolicy(props) {
                                 smooth={true}
                                 duration={1000}
                             >
-                                <li className="side-menu__item">Email communication</li>
+                                <li className="side-menu__item">Responsibilities</li>
                             </Link>
                             <Link
                                 activeClass='item-active'
@@ -56,7 +53,7 @@ function PrivacyPolicy(props) {
                                 smooth={true}
                                 duration={1000}
                             >
-                                <li className="side-menu__item">Public Profile</li>
+                                <li className="side-menu__item">Your content</li>
                             </Link>
                             <Link
                                 activeClass='item-active'
@@ -65,29 +62,28 @@ function PrivacyPolicy(props) {
                                 smooth={true}
                                 duration={1000}
                             >
-                            <li className="side-menu__item">Log files</li>
+                                <li className="side-menu__item">Licensing to Mango</li>
                             </Link>
                         </ul>
                     </div>
-                    <div className='privacy-policy-text'>
-                        <h2 className='sub-header' id='section1'>Privacy Statement and Notice</h2>
+                    <div className='terms-of-use__text'>
+                        <h2 className='terms-of-use__sub-header' id='section1'>Introduction</h2>
                         <p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.</p>
-                        <h2 className='sub-header' id='section2'>What information we collect</h2>
+                        <h2 className='terms-of-use__sub-header' id='section2'>Services</h2>
                         <p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.</p>
-                        <h2 className='sub-header' id='section3'>Sharing your information</h2>
+                        <h2 className='terms-of-use__sub-header' id='section3'>Registration</h2>
                         <p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.</p>
-                        <h2 className='sub-header' id='section4'>Email communication</h2>
+                        <h2 className='terms-of-use__sub-header' id='section4'>Responsibilities</h2>
                         <p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.</p>
-                        <h2 className='sub-header' id='section5'>Public profile</h2>
+                        <h2 className='terms-of-use__sub-header' id='section5'>Your content</h2>
                         <p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.</p>
-                        <h2 className='sub-header' id='section6'>Log files</h2>
+                        <h2 className='terms-of-use__sub-header' id='section6'>Licensing to Mango</h2>
                         <p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.</p>
                     </div>
                 </div>
             </div>
-            <Footer/>
         </div>
     )
 }
 
-export default PrivacyPolicy;
+export default TermsOfUse;
