+++ frontend/src/components/Main/TermsOfService/TermsOfUse.scss
@@ -1,36 +1,37 @@
-@import '../../styles/_variables.scss';
+@import '../../../styles/variables';
 
-.privacy-policy{
+.terms-of-use {
+  padding-top: $MAIN_MENU_HEIGHT;
 
-  .privacy-policy-container {
+  &__container {
     width: 1200px;
-    min-height: calc(100vh - #{$MAIN_MENU_HEIGHT} - #{$FOOTER_HEIGHT});
+    min-height: calc(100vh - #{$FOOTER_HEIGHT});
     margin: 0 auto;
     padding: 0 30px;
-
-    &__header {
-      font-size: 32px;
-      font-weight: 600;
-      color: $COLOR_MENU_DARK;
-      text-transform: uppercase;
-      padding: 50px 0;
-    }
   }
 
-  .privacy-policy-content {
+  &__header {
+    font-size: 32px;
+    font-weight: 600;
+    color: $COLOR_MENU_DARK;
+    text-transform: uppercase;
+    padding: 50px 0;
+  }
+
+  &__text {
+    width: 70%;
+  }
+
+  &__sub-header {
+    font-size: 28px;
+    color: #333333;
+    margin-bottom: 10px;
+  }
+
+  &__content {
     display: flex;
     font-size: 16px;
     line-height: 21px;
-
-    .privacy-policy-text {
-      width: 70%;
-
-      .sub-header {
-        font-size: 28px;
-        color: #333333;
-        margin-bottom: 10px;
-      }
-    }
 
     p {
       margin-bottom: 30px;
@@ -41,14 +42,15 @@
     }
 
     .side-menu {
-    width: 30%;
+      width: 30%;
 
       &__item {
         font-size: 18px;
         cursor: pointer;
+        color: #333333;
       }
 
-      &__item:hover:after{
+      &__item:hover:after {
         background: rgba(54, 179, 168, 0.9);
         height: 2px;
         margin-bottom: 10px;
@@ -63,7 +65,7 @@
         margin: 11px 0;
         background-color: #BDBDBD;
       }
-  }
+    }
   }
 
   .item-active > .side-menu__item::after {
