+++ frontend/src/containers/Main/MainPage.js
@@ -6,7 +6,6 @@ import {Route, Switch} from 'react-router-dom';
 import routes from '../../constants/routes';
 import ForTalents from '../../components/Main/ForTalents/ForTalents';
 import ForEmployers from '../../components/Main/ForEmployers/ForEmployers';
-import BlogPage from './Blog/BlogPage';
 import Home from '../../components/Main/Home/Home';
 import Footer from '../../components/Main/Footer/Footer';
 import './Main.scss';
@@ -14,6 +13,10 @@ import About from '../../components/Main/About/About';
 import JobsRolesSelect
   from '../../components/Main/Jobs/JobsRolesSelect/JobsRoleSelect';
 import JobsByRole from '../../components/Main/Jobs/JobsByRole/JobsByRole';
+import PrivacyPolicy from '../../components/Main/PrivacyPolicy/PrivacyPolicy';
+import TermsOfUse from '../../components/Main/TermsOfService/TermsOfUse';
+import Support from '../../components/Main/Support/Support';
+import Blog from '../../components/Main/Blog/Blog';
 
 type Props = {
   history: Object
@@ -32,8 +35,11 @@ class MainPage extends Component<Props> {
               <Route exact path={routes.FIND_JOB} component={JobsRolesSelect}/>
               <Route path={routes.FIND_JOB + '/:jobRole'}
                      component={JobsByRole} url={routes.FIND_JOB}/>
-              <Route path={routes.BLOG} component={BlogPage}/>
               <Route path={routes.ABOUT} component={About}/>
+              <Route path={routes.PRIVACY_POLICY} component={PrivacyPolicy}/>
+              <Route path={routes.TERMS_OF_USE} component={TermsOfUse}/>
+              <Route path={routes.SUPPORT_CENTER} component={Support}/>
+              <Route path={routes.BLOG} component={Blog}/>
               <Route path={'/*'} component={()=>(<div>404</div>)}/>
             </Switch>
           </div>
