--- frontend/src/pages/Blog/Blog.css
@@ -1,93 +1,0 @@
-.blog-header {
-  width: 100%;
-  height: 500px;
-  background-image: url("../../assets/images/bg-blog.png");
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: cover;
-  position: relative; }
-
-.blog-header__title {
-  width: 988px;
-  padding: 0 10px;
-  font-size: 48px;
-  font-weight: bold;
-  color: #fff;
-  z-index: 99999;
-  position: absolute;
-  text-transform: uppercase;
-  text-align: center;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  line-height: 50px; }
-
-.blog-header__image-filter {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: #060C2E;
-  opacity: 50%; }
-
-.blog-container {
-  width: 988px;
-  margin: 0 auto;
-  padding: 0 10px;
-  display: flex;
-  margin-top: 87px; }
-
-.blog-container .blog-content {
-  width: 70%;
-  margin-right: 70px; }
-
-.blog-container .blog-item {
-  color: #828282;
-  margin-bottom: 67px; }
-
-.blog-container .blog-item__description {
-  margin-bottom: 4px;
-  font-size: 12px;
-  text-transform: uppercase; }
-
-.blog-container .blog-item__container {
-  display: flex;
-  margin-bottom: 16px; }
-
-.blog-container .blog-item__header {
-  margin-bottom: 25px;
-  color: #4F4F4F;
-  font-size: 36px;
-  font-weight: 600; }
-
-.blog-container .blog-item__text {
-  font-size: 24px;
-  line-height: 29px; }
-
-.blog-container .blog-item__icon {
-  margin-right: 24px;
-  width: 180px;
-  height: 115px;
-  flex-shrink: 0;
-  background-size: cover;
-  background-repeat: no-repeat;
-  background-position: center;
-  background-image: url("../../assets/images/rachel-moenning-570271-unsplash.png"); }
-
-.blog-container .blog-item__link {
-  text-transform: uppercase;
-  text-decoration: underline;
-  font-size: 18px;
-  cursor: pointer;
-  color: #828282; }
-
-.blog-container .blog-item__link:hover {
-  text-decoration: none; }
-
-.blog-container .search-form .search-input {
-  width: 306px;
-  height: 44px;
-  box-sizing: border-box;
-  padding-left: 44px;
-  border: 1px solid #BDBDBD; }
